Function to set data in the correct input format for model-averaging functions, for continuous data with covariate effect (used internally)

The function takes a dataset as input and generates the data list, starting values, and prior distributions needed by the stan models to be fitted.

PREP_DATA_NCOV(
  data,
  sumstats = TRUE,
  geom.stats = FALSE,
  sd = TRUE,
  q,
  bkg = NULL,
  maxy = NULL,
  prior.BMD = NULL,
  shape.a = 4,
  shape.c = 4,
  shape.BMD = 1e-04,
  prior.d = "N11",
  d.mean = NULL,
  d.std = NULL,
  d.trunc = NULL,
  extended = TRUE,
  extended.value = 3,
  covariate = "all"
)

PREP_DATA_LNCOV(
  data,
  sumstats = TRUE,
  geom.stats = FALSE,
  sd = TRUE,
  q,
  bkg = NULL,
  maxy = NULL,
  prior.BMD = NULL,
  shape.a = 4,
  shape.c = 4,
  shape.BMD = 1e-04,
  prior.d = "N11",
  d.mean = NULL,
  d.std = NULL,
  d.trunc = NULL,
  extended = TRUE,
  extended.value = 3,
  covariate = "all"
)

Arguments

data

a dataframe with input data, order of columns should be: dose, response, SD or SE, n, covariate level

sumstats

logical indicating whether summary (T, default) or individual-level (F) data is provided

geom.stats

logicial indicating whether, if summary data are provided, these are geometric (T) or arithmetic (F, default) summary statistics

sd

logical indicating whether standard deviation (T, default) or standard error (F) is provided

q

specified BMR

bkg

vector containing minimum, most likely (optional, can be NA), and maximum value for the background response. Defaults to NULL (non-informative prior)

maxy

vector containing minimum, most likely (optional, can be NA), and maximum value for the response at dose infinity. Defaults to NULL (non-informative prior)

prior.BMD

vector containing minimum, most likely (optional, can be NA), and maximum value for the BMD. Defaults to NULL (non-informative prior)

shape.a

shape parameter for the modified PERT distribution on parameter a, defaults to 4 (peaked at most likely value), a value of 0.0001 implies a uniform distribution

shape.c

shape parameter for the modified PERT distribution on parameter c, defaults to 4 (peaked at most likely value), a value of 0.0001 implies a uniform distribution

shape.BMD

shape parameter for the modified PERT distribution on parameter BMD, defaults to 0.0001 implying a uniform distribution. Can be set to 4 in case of informative prior

prior.d

prior distribution for parameter d (on log scale), should be either N11 (default N(1, 1) prior truncated at 5), EPA (N(0.4, sqrt(0.5)) prior) or N05 (for a N(0.5,0.5) prior)

d.mean

mean of the Normal prior for log(d), only used when prior.d = 'custom'

d.std

sd of the Normal prior for log(d), only used when prior.d = 'custom'

d.trunc

truncation (upper) of the Normal prior for log(d), only used when prior.d = 'custom'

extended

logical indicating whether the dose range should be extended to maxDose^2 (default is TRUE)

extended.value

value for the upper limit of BMD prior in extended range

covariate

on which parameters a covariate effect should be used. Defaults to 'all', other options are 'a_sigma2' and 'BMD_d'

Value

List with data and start values in correct format to be directly used within the model-averaging functions.