Function to set data in the correct input format for model-averaging functions, for clustered continuous data

This function takes a dataset as input and generates the data list, starting values, and prior distributions needed by the stan models to be fitted. Shape parameters for the prior distribution can be any value, with 0.0001 resulting in a uniform prior and 4 resulting in a peaked prior. For example, setting the shape parameter to 2 results in a slightly peaked prior distribution.

PREP_DATA_N_C(
  data,
  q,
  bkg = NULL,
  maxy = NULL,
  prior.BMD = NULL,
  shape.a = 4,
  shape.c = 4,
  shape.BMD = 1e-04,
  prior.d = "N11",
  d.mean = NULL,
  d.std = NULL,
  d.trunc = NULL,
  extended = T,
  extended.value = 3
)

PREP_DATA_LN_C(
  data,
  q,
  bkg = NULL,
  maxy = NULL,
  prior.BMD = NULL,
  shape.a = 4,
  shape.c = 4,
  shape.BMD = 1e-04,
  prior.d = "N11",
  d.mean = NULL,
  d.std = NULL,
  d.trunc = NULL,
  extended = T,
  extended.value = 3
)

Arguments

data

a dataframe with individual-level input data, order of columns should be: dose, response, litter

q

specified BMR

bkg

vector containing minimum, most likely (optional, can be NA), and maximum value for the background response. Defaults to NULL (non-informative prior)

maxy

vector containing minimum, most likely (optional, can be NA), and maximum value for the response at dose infinity. Defaults to NULL (non-informative prior)

prior.BMD

vector containing minimum, most likely (optional, can be NA), and maximum value for the BMD. Defaults to NULL (non-informative prior)

shape.a

shape parameter for the modified PERT distribution on parameter a, defaults to 4 (peaked at most likely value), a value of 0.0001 implies a uniform distribution

shape.c

shape parameter for the modified PERT distribution on parameter c, defaults to 4 (peaked at most likely value), a value of 0.0001 implies a uniform distribution

shape.BMD

shape parameter for the modified PERT distribution on parameter BMD, defaults to 0.0001 implying a uniform distribution. Can be set to 4 in case of informative prior

prior.d

prior distribution for parameter d (on log scale), should be either N11 (default N(1, 1) prior truncated at 5), EPA (N(0.4, sqrt(0.5)) prior), N05 (for a N(0.5,0.5) prior), or 'custom'

d.mean

mean of the Normal prior for log(d), only used when prior.d = 'custom'

d.std

sd of the Normal prior for log(d), only used when prior.d = 'custom'

d.trunc

truncation (upper) of the Normal prior for log(d), only used when prior.d = 'custom'

extended

logical indicating whether the dose range should be extended to maxDose^2 (default is TRUE)

extended.value

value for the upper limit of BMD prior in extended range

Value

data list containing the data and prior distributions

start list containing start values for the model parameters

startQ list containing start values for the quadratic exponential model

shapiro.p p-value of the Shapiro-Wilks test for normality

bartlett.p p-value of the Bartlett test for homoskedasticity

shapiro.msg message with result of test for normality

bartlett.msg message with result of test for homoskedasticity

Examples

data.input <- das5$data
data.input <- data.frame(dose = data.input$dose, response = data.input$foetalweight, litter = data.input$number)

# default priors
data_N <- PREP_DATA_N_C(data.input, q = 0.05, prior.d = 'N11')
#> GAMLSS-RS iteration 1: Global Deviance = -29.0541 
#> GAMLSS-RS iteration 2: Global Deviance = -29.0541 
#> Warning: the data do not contain information on the asymptote, and the default prior for fold change has been based on half the observed maximum;
#>             please provide prior input on the maximum response by specifying 'maxy', if available
#> default prior choices used on background
#> default prior choices used on fold change
#> default prior choices used on BMD
#> GAMLSS-RS iteration 1: Global Deviance = 1984.496 
#> GAMLSS-RS iteration 2: Global Deviance = 1984.496 
#> Warning: The origing of the x variable has been shifted by 1e-14
#> Warning: The origing of the x variable has been shifted by 1e-14
#> Warning: distributional assumption of constant variance for the normal distribution is not met, Bartlett test p-value is 0.0009
#> Warning: distributional assumption of normality of residuals for the normal distribution is not met, Shapiro test p-value is 0.0000