Model averaging using Full Laplace method

This function performs model averaging using the full Laplace approximation. By default, all 16 models are included for continuous data, and all 8 models for quantal data. Models can be excluded by setting their respective weight to 0 in prior.weights. The order of models fitted can be obtained using the get_models() function.

full.laplace_MA is used for continuous data

full.laplace_MAc is used for clustered continuous data (i.e. with litter effect)

full.laplaceQ_MA is used for quantal data (with or without litter effect)

full.laplace_MA(
  data.N,
  data.LN,
  prior.weights = rep(1, 16),
  ndraws = 30000,
  seed = 123,
  pvec = c(0.05, 0.5, 0.95),
  plot = FALSE
)

full.laplace_MAc(
  data.N,
  data.LN,
  prior.weights = rep(1, 16),
  ndraws = 30000,
  seed = 123,
  pvec = c(0.05, 0.5, 0.95)
)

full.laplaceQ_MA(
  data.Q,
  prior.weights = rep(1, 8),
  ndraws = 30000,
  seed = 123,
  pvec = c(0.05, 0.5, 0.95)
)

Arguments

data.N

the input data as returned by function PREP_DATA_N or PREP_DATA_N_C for clustered data

data.LN

the input data as returned by function PREP_DATA_LN or PREP_DATA_LN_C for clustered data

prior.weights

a vector specifying which of the 16 (continuous) or 8 (quantal) models should be included (1 = include, 0 = exclude)

ndraws

the number of draws, default 30000

seed

random seed, default 123

pvec

vector specifying the three BMD quantiles of interest (default 90% CrI)

data.Q

the input data as returned by function PREP_DATA_Q

Value

The function returns a BMDBMA model object, which is a list containing the following objects:

E4_N parameter estimates from the (Exponential, Normal) model

IE4_N parameter estimates from the (Inverse Exponential, Normal) model

H4_N parameter estimates from the (Hill, Normal) model

LN4_N parameter estimates from the (Log-normal, Normal)

G4_N parameter estimates from the (Gamma, Normal) model

QE4_N parameter estimates from the (Quadratic Exponential, Normal) model

P4_N parameter estimates from the (Probit, Normal) model

L4_N parameter estimates from the (Logit, Normal) model

E4_LN parameter estimates from the (Exponential, Log-normal) model

IE4_LN parameter estimates from the (Inverse Exponential, Log-normal) model

H4_LN parameter estimates from the (Hill, Log-normal) model

LN4_LN parameter estimates from the (Log-normal, Log-normal)

G4_LN parameter estimates from the (Gamma, Log-normal) model

QE4_LN parameter estimates from the (Quadratic Exponential, Log-normal) model

P4_LN parameter estimates from the (Probit, Log-normal) model

L4_LN parameter estimates from the (Logit, Log-normal) model

E4_Q parameter estimates from the (Exponential, Quantal) model

IE4_Q parameter estimates from the (Inverse Exponential, Quantal) model

H4_Q parameter estimates from the (Hill, Quantal) model

LN4_Q parameter estimates from the (Lognormal, Quantal) model

G4_Q parameter estimates from the (Gamma, Quantal) model

QE4_Q parameter estimates from the (Quadratic Exponential, Quantal) model

P4_Q parameter estimates from the (Probit, Quantal) model

L4_Q parameter estimates from the (Logit, Quantal) model

MA Model averaged BMD credible interval

weights Model weights used in the averaging

bf Bayes factor comparing the best model against saturated ANOVA model

covs matrix with covariances between parameters b-d and BMD-d

corrs matrix with correlation between parameters b-d and BMD-d

p.msg warning message if model averaged posterior has been truncated

w.msg warning message if Laplace weights could not be computed and one model gets all the weight

shift shift value for lognormal data

BIC.SM BIC value of saturated model, used to test for GOF

BIC.bestfit BIC value of best fitting model, used to test for GOF

means.SM mean response per dose level estimated from the saturated model, used to test for GOF

gof_check GOF message

models_included_laplace vector containing the names of models included in the model averaging

q BMR

max.dose maximum dose level (original scale)

dataN normal summary data used for analysis

dataLN lognormal summary data used for analysis

BMDMixture vector of length ndraws containing the draws from the model-averaged posterior

MA_dr vector containing the model-averaged response at each dose level

MA_post vector containing the 0.5%-percentiles of the model-averaged posterior

llN vector containing the loglikelihood values of the Normal models

llLN vector containing the loglikelihood values of the Lognormal models

parsN list containing the fitted Normal models

parsLN list containing the fitted Lognormal models

is_bin logical indicating whether binomial (no litter) model was used

is_betabin logical indicating whether betabinomial (with litter) model was used

data quantal summary data used for analysis

parsQ list containing the fitted Quantal models

llQ vector containing the loglikelihood values of the Quantal models

Examples

 data_N <- PREP_DATA_N(data = as.data.frame(immunotoxicityData[1:5,]),
                       sumstats = TRUE, sd = TRUE, q = 0.1)
#> distributional assumption of constant variance is met, Bartlett test p-value is 0.0527
#> GAMLSS-RS iteration 1: Global Deviance = -35.5626 
#> GAMLSS-RS iteration 2: Global Deviance = -35.5626 
#> Warning: the data do not contain information on the asymptote, and the default prior for fold change has been based on 3 times the observed maximum;
#>             please provide prior input on the maximum response by specifying 'maxy', if available
#> default prior choices used on background
#> default prior choices used on fold change
#> default prior choices used on BMD
#> GAMLSS-RS iteration 1: Global Deviance = -38.2921 
#> GAMLSS-RS iteration 2: Global Deviance = -38.2921 
#> Warning: The origing of the x variable has been shifted by 1e-14
#> Warning: The origing of the x variable has been shifted by 1e-14
 data_LN <- PREP_DATA_LN(data = as.data.frame(immunotoxicityData[1:5,]),
                         sumstats = TRUE, sd = TRUE, q = 0.1) #'
#> distributional assumption of constant variance (on log-scale) is met, Bartlett test p-value is 0.6799
#> GAMLSS-RS iteration 1: Global Deviance = -35.5626 
#> GAMLSS-RS iteration 2: Global Deviance = -35.5626 
#> Warning: the data do not contain information on the asymptote, and the default prior for fold change has been based on 3 times the observed maximum;
#>             please provide prior input on the maximum response by specifying 'maxy', if available
#> default prior choices used on background
#> default prior choices used on fold change
#> default prior choices used on BMD
#> GAMLSS-RS iteration 1: Global Deviance = -38.2921 
#> GAMLSS-RS iteration 2: Global Deviance = -38.2921 
#> Warning: The origing of the x variable has been shifted by 1e-14
#> Warning: The origing of the x variable has been shifted by 1e-14
 FLBMD <- full.laplace_MA(data_N, data_LN, prior.weights = c(rep(1,4), rep(0,12)))
#> Warning: BMDU/BMDL is larger than 50
#> Warning: Best fitting model fits sufficiently well (Bayes factor is 1.52e-01).