Perform model averaging using Full Laplace method for data with covariate effect

This function performs model averaging using the full Laplace approximation. By default, all 16 models are included for continuous data, and all 8 models for quantal data. Models can be excluded by setting their respective weight to 0 in prior.weights. The order of models fitted can be obtained using the get_models() function.

full.laplace_MA_Cov is used for continuous data

full.laplace_MA_Q_Cov is used for quantal data

full.laplace_MA_Cov(
  data,
  sumstats = TRUE,
  sd = TRUE,
  q = 0.05,
  prior.d = "N11",
  extended = TRUE,
  extended.value = 3,
  prior.weights = rep(1, 16),
  ndraws = 30000,
  seed = 123,
  pvec = c(0.05, 0.5, 0.95)
)

full.laplace_MA_Q_Cov(
  data,
  sumstats = TRUE,
  q = 0.1,
  prior.d = "N11",
  extended = TRUE,
  extended.value = 3,
  prior.weights = rep(1, 8),
  ndraws = 30000,
  seed = 123,
  pvec = c(0.05, 0.5, 0.95)
)

Arguments

data

the summary data, with columns: dose, response, sd, n, covariate level for continuous; dose, number of adverse events, n, covariate for quantal

sumstats

logical indicating whether summary (T, default) or individual-level (F) data is provided

sd

logical indicating whether standard deviation (T, default) or standard error (F) is provided

q

specified BMR

prior.d

prior distribution for parameter d (on log scale), should be either N11 (default N(1, 1) prior truncated at 5), EPA (N(0.4, sqrt(0.5)) prior) or N05 (for a N(0.5,0.5) prior)

extended

logical indicating whether the dose range should be extended to maxDose^2 (default is TRUE)

extended.value

value for the upper range of BMD prior

prior.weights

a vector specifying which of the 16 (continuous) or 8 (quantal) models should be included (1 = include, 0 = exclude)

ndraws

the number of draws, default 30000

seed

default 123

pvec

vector specifying the three BMD quantiles of interest (default 90% CrI)

Value

MA Model averaged BMD credible interval per covariate level (in case of an effect)

weights Weights for the best fitting submodels used in model averaging

summary Summary table of results

data Data used in analysis

maxDose Maximum dose level (original scale)

BMD_Mixture 0.5%-percentiles of the model averaged posterior for the BMD

parE4_N Estimated parameters for the (Exponential, Normal) model

parIE4_N Estimated parameters for the (Inverse Exponential, Normal) model

parH4_N Estimated parameters for the (Hill, Normal) model

parLN4_N Estimated parameters for the (Lognormal, Normal) model

parG4_N Estimated parameters for the (Gamma, Normal) model

parQE4_N Estimated parameters for the (Quadratic Exponential, Normal) model

parP4_N Estimated parameters for the (Probit, Normal) model

parL4_N Estimated parameters for the (Logit, Normal) model

parE4_LN Estimated parameters for the (Exponential, Lognormal) model

parIE4_LN Estimated parameters for the (Inverse Exponential, Lognormal) model

parH4_LN Estimated parameters for the (Hill, Lognormal) model

parLN4_LN Estimated parameters for the (Lognormal, Lognormal) model

parG4_LN Estimated parameters for the (Gamma, Lognormal) model

parQE4_LN Estimated parameters for the (Quadratic Exponential, Lognormal) model

parP4_LN Estimated parameters for the (Probit, Lognormal) model

parL4_LN Estimated parameters for the (Logit, Lognormal) model

parE4_Q Estimated parameters for the (Exponential, Quantal) model

parIE4_Q Estimated parameters for the (Inverse Exponential, Quantal) model

parH4_Q Estimated parameters for the (Hill, Quantal) model

parLN4_Q Estimated parameters for the (Lognormal, Quantal) model

parG4_Q Estimated parameters for the (Gamma, Quantal) model

parQE4_Q Estimated parameters for the (Quadratic Exponential, Quantal) model

parP4_Q Estimated parameters for the (Probit, Quantal) model

parL4_Q Estimated parameters for the (Logit, Quantal) model

shift shift in case of negative geometric means

q BMR

Examples

summ.data = data.frame(x = data_cont_covariate$Dose, y = data_cont_covariate$Response, s = data_cont_covariate$SD, n = data_cont_covariate$N, cov = data_cont_covariate$Covariate)
FLBMD <- full.laplace_MA_Cov(summ.data, sumstats = T, sd = T, q = 0.05, prior.d = 'N11', extended = T, ndraws = 30000, seed = 123, pvec = c(0.05, 0.5, 0.95), prior.weights = rep(1,16))
#> distributional assumption of constant variance are met for group Male, Bartlett test p-value is 0.7627
#> distributional assumption of constant variance are met for group Female, Bartlett test p-value is 0.5937
#> GAMLSS-RS iteration 1: Global Deviance = -20.379 
#> GAMLSS-RS iteration 2: Global Deviance = -20.379 
#> default prior choices used on background
#> default prior choices used on fold change
#> GAMLSS-RS iteration 1: Global Deviance = -20.379 
#> GAMLSS-RS iteration 2: Global Deviance = -20.379 
#> default prior choices used on background
#> default prior choices used on fold change
#> GAMLSS-RS iteration 1: Global Deviance = -26.3218 
#> GAMLSS-RS iteration 2: Global Deviance = -26.3218 
#> Warning: The origing of the x variable has been shifted by 1e-14
#> Warning: The origing of the x variable has been shifted by 1e-14
#> GAMLSS-RS iteration 1: Global Deviance = -25.1272 
#> GAMLSS-RS iteration 2: Global Deviance = -25.1272 
#> Warning: The origing of the x variable has been shifted by 1e-14
#> Warning: The origing of the x variable has been shifted by 1e-14
#> default prior choices used on BMD
#> distributional assumption of constant variance is met, Bartlett test p-value is 0.8792
#> GAMLSS-RS iteration 1: Global Deviance = -20.379 
#> GAMLSS-RS iteration 2: Global Deviance = -20.379 
#> Warning: the data do not contain information on the asymptote, and the default prior for fold change has been based on half the observed maximum;
#>             please provide prior input on the maximum response by specifying 'maxy', if available
#> default prior choices used on background
#> default prior choices used on fold change
#> default prior choices used on BMD
#> GAMLSS-RS iteration 1: Global Deviance = 26.4866 
#> GAMLSS-RS iteration 2: Global Deviance = 26.4866 
#> Warning: The origing of the x variable has been shifted by 1e-14
#> Warning: The origing of the x variable has been shifted by 1e-14
#> distributional assumption of constant variance are met for group Male, Bartlett test p-value is 0.7627
#> distributional assumption of constant variance are met for group Female, Bartlett test p-value is 0.5937
#> GAMLSS-RS iteration 1: Global Deviance = -20.379 
#> GAMLSS-RS iteration 2: Global Deviance = -20.379 
#> default prior choices used on background
#> default prior choices used on fold change
#> GAMLSS-RS iteration 1: Global Deviance = -20.379 
#> GAMLSS-RS iteration 2: Global Deviance = -20.379 
#> default prior choices used on background
#> default prior choices used on fold change
#> GAMLSS-RS iteration 1: Global Deviance = -18.0344 
#> GAMLSS-RS iteration 2: Global Deviance = -18.0344 
#> Warning: The origing of the x variable has been shifted by 1e-14
#> Warning: The origing of the x variable has been shifted by 1e-14
#> default prior choices used on BMD
#> distributional assumption of constant variance is met, Bartlett test p-value is 0.3881
#> GAMLSS-RS iteration 1: Global Deviance = -33.767 
#> GAMLSS-RS iteration 2: Global Deviance = -33.767 
#> default prior choices used on background
#> default prior choices used on fold change
#> GAMLSS-RS iteration 1: Global Deviance = -26.3218 
#> GAMLSS-RS iteration 2: Global Deviance = -26.3218 
#> Warning: The origing of the x variable has been shifted by 1e-14
#> Warning: The origing of the x variable has been shifted by 1e-14
#> GAMLSS-RS iteration 1: Global Deviance = -25.1272 
#> GAMLSS-RS iteration 2: Global Deviance = -25.1272 
#> Warning: The origing of the x variable has been shifted by 1e-14
#> Warning: The origing of the x variable has been shifted by 1e-14
#> default prior choices used on BMD
#> distributional assumption of constant coefficient of variation is met for group Male, Bartlett test p-value is 0.5923 
#> distributional assumption of constant coefficient of variation is met for group Female, Bartlett test p-value is 0.2813 
#> GAMLSS-RS iteration 1: Global Deviance = -20.379 
#> GAMLSS-RS iteration 2: Global Deviance = -20.379 
#> default prior choices used on background
#> default prior choices used on fold change
#> GAMLSS-RS iteration 1: Global Deviance = -20.379 
#> GAMLSS-RS iteration 2: Global Deviance = -20.379 
#> default prior choices used on background
#> default prior choices used on fold change
#> GAMLSS-RS iteration 1: Global Deviance = -26.3218 
#> GAMLSS-RS iteration 2: Global Deviance = -26.3218 
#> Warning: The origing of the x variable has been shifted by 1e-14
#> Warning: The origing of the x variable has been shifted by 1e-14
#> GAMLSS-RS iteration 1: Global Deviance = -25.1272 
#> GAMLSS-RS iteration 2: Global Deviance = -25.1272 
#> Warning: The origing of the x variable has been shifted by 1e-14
#> Warning: The origing of the x variable has been shifted by 1e-14
#> default prior choices used on BMD
#> distributional assumption of constant coefficient of variation is met for group Male, Bartlett test p-value is 0.5923 
#> distributional assumption of constant coefficient of variation is met for group Female, Bartlett test p-value is 0.2813 
#> GAMLSS-RS iteration 1: Global Deviance = -20.379 
#> GAMLSS-RS iteration 2: Global Deviance = -20.379 
#> default prior choices used on background
#> default prior choices used on fold change
#> GAMLSS-RS iteration 1: Global Deviance = -20.379 
#> GAMLSS-RS iteration 2: Global Deviance = -20.379 
#> default prior choices used on background
#> default prior choices used on fold change
#> GAMLSS-RS iteration 1: Global Deviance = -18.0344 
#> GAMLSS-RS iteration 2: Global Deviance = -18.0344 
#> Warning: The origing of the x variable has been shifted by 1e-14
#> Warning: The origing of the x variable has been shifted by 1e-14
#> default prior choices used on BMD
#> distributional assumption of constant coefficient of variation is met, Bartlett test p-value is %1$5.4f0.1316
#> GAMLSS-RS iteration 1: Global Deviance = -33.767 
#> GAMLSS-RS iteration 2: Global Deviance = -33.767 
#> default prior choices used on background
#> default prior choices used on fold change
#> GAMLSS-RS iteration 1: Global Deviance = -26.3218 
#> GAMLSS-RS iteration 2: Global Deviance = -26.3218 
#> Warning: The origing of the x variable has been shifted by 1e-14
#> Warning: The origing of the x variable has been shifted by 1e-14
#> GAMLSS-RS iteration 1: Global Deviance = -25.1272 
#> GAMLSS-RS iteration 2: Global Deviance = -25.1272 
#> Warning: The origing of the x variable has been shifted by 1e-14
#> Warning: The origing of the x variable has been shifted by 1e-14
#> default prior choices used on BMD
#> distributional assumption of constant variance (on log-scale) is met, Bartlett test p-value is 0.4963
#> GAMLSS-RS iteration 1: Global Deviance = -20.379 
#> GAMLSS-RS iteration 2: Global Deviance = -20.379 
#> Warning: the data do not contain information on the asymptote, and the default prior for fold change has been has been based on half the observed maximum;
#>             please provide prior input on the maximum response by specifying 'maxy', if available
#> default prior choices used on background
#> default prior choices used on fold change
#> default prior choices used on BMD
#> GAMLSS-RS iteration 1: Global Deviance = 26.4866 
#> GAMLSS-RS iteration 2: Global Deviance = 26.4866 
#> Warning: The origing of the x variable has been shifted by 1e-14
#> Warning: The origing of the x variable has been shifted by 1e-14
#> Error in chol.default(-H) : the leading minor of order 7 is not positive