Function used internally to test whether the best fitting model fits the data well, compared to a saturated (non-monotone) model

The function compares a dose-response model to the saturated model using an approximation to BIC.

modelTest(
  best.fit,
  data.N,
  data.LN,
  stanBest,
  type,
  seed,
  ndraws,
  nrchains,
  nriterations,
  warmup,
  delta,
  treedepth
)

modelTestC(
  best.fit,
  data.N,
  data.LN,
  stanBest,
  type,
  seed,
  ndraws,
  nrchains,
  nriterations,
  warmup,
  delta,
  treedepth
)

modelTestQ(
  best.fit,
  data.Q,
  stanBest,
  type,
  seed,
  ndraws,
  nrchains,
  nriterations,
  warmup,
  delta,
  treedepth
)

Arguments

best.fit

model name

data.N

object as given by PREP_DATA_N

data.LN

object as given by PREP_DATA_LN

stanBest

stan object for the model given in best.fit

type

estimation type. Laplace or MCMC

seed

random seed for reproducibility

ndraws

number of draws to be made from the posterior distribution. Defaults to 30000

nrchains

number of MCMC chains. Defaults to 3

nriterations

number of MCMC iterations.Defaults to 3000

warmup

number of MCMC iterations for warmup. Defaults to 1000

delta

adapt_delta value for the HMC in stan. See sampling for more. Defaults to 0.8.

treedepth

tree_depth value for the HMC in stan. See sampling for more. Defaults to 10.

data.Q

object as given by PREP_DATA_Q

Value

Bayes factor for the best fitting model compared to the saturated ANOVA model.